+++ src/test/java/za/ac/sun/cs/coastal/SystemTestsArrays.java
@@ -19,7 +19,7 @@ public class SystemTestsArrays extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 3);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 3, 0, 0);
 		checkInstrumentation(reporter, 15, 4, 1);
 	}
@@ -34,7 +34,7 @@ public class SystemTestsArrays extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 3);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 3, 0, 0);
 		checkInstrumentation(reporter, 24, 4, 1);
 	}
@@ -49,7 +49,7 @@ public class SystemTestsArrays extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 3);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 3, 0, 0);
 		checkInstrumentation(reporter, 15, 4, 1);
 	}
@@ -64,7 +64,7 @@ public class SystemTestsArrays extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 3);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 3, 0, 0);
 		checkInstrumentation(reporter, 15, 4, 1);
 	}
@@ -79,7 +79,7 @@ public class SystemTestsArrays extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 3);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 3, 0, 0);
 		checkInstrumentation(reporter, 15, 4, 1);
 	}
@@ -94,7 +94,7 @@ public class SystemTestsArrays extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 3);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 3, 0, 0);
 		checkInstrumentation(reporter, 15, 4, 1);
 	}