+++ src/test/java/za/ac/sun/cs/coastal/SystemTestsEntryPoint.java
@@ -9,7 +9,7 @@ public class SystemTestsEntryPoint extends SystemTests {
 		Reporter reporter = setup("tests/Test01.xml", "tests/entrypoint/EntryPoint01-A.xml");
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 20, 6, 1);
 		checkMarkerCoverage(reporter, 0, 0, 1);
@@ -21,7 +21,7 @@ public class SystemTestsEntryPoint extends SystemTests {
 		Reporter reporter = setup("tests/Test01.xml", "tests/entrypoint/EntryPoint01-B.xml");
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 20, 6, 1);
 		checkMarkerCoverage(reporter, 0, 0, 1);
@@ -33,7 +33,7 @@ public class SystemTestsEntryPoint extends SystemTests {
 		Reporter reporter = setup("tests/Test01.xml", "tests/entrypoint/EntryPoint01-C.xml");
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 20, 6, 1);
 		checkMarkerCoverage(reporter, 0, 0, 1);
@@ -45,7 +45,7 @@ public class SystemTestsEntryPoint extends SystemTests {
 		Reporter reporter = setup("tests/Test01.xml", "tests/entrypoint/EntryPoint01-D.xml");
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 20, 6, 1);
 		checkMarkerCoverage(reporter, 0, 0, 1);
@@ -57,7 +57,7 @@ public class SystemTestsEntryPoint extends SystemTests {
 		Reporter reporter = setup("tests/Test01.xml", "tests/entrypoint/EntryPoint01-E.xml");
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 20, 6, 1);
 		checkMarkerCoverage(reporter, 0, 0, 1);
@@ -69,7 +69,7 @@ public class SystemTestsEntryPoint extends SystemTests {
 		Reporter reporter = setup("tests/Test01.xml", "tests/entrypoint/EntryPoint01-F.xml");
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 20, 6, 1);
 		checkMarkerCoverage(reporter, 0, 0, 1);
@@ -81,7 +81,7 @@ public class SystemTestsEntryPoint extends SystemTests {
 		Reporter reporter = setup("tests/Test01.xml", "tests/entrypoint/EntryPoint01-G.xml");
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 20, 6, 1);
 		checkMarkerCoverage(reporter, 0, 0, 1);
@@ -93,7 +93,7 @@ public class SystemTestsEntryPoint extends SystemTests {
 		Reporter reporter = setup("tests/Test01.xml", "tests/entrypoint/EntryPoint02-A.xml");
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 24, 6, 1);
 		checkMarkerCoverage(reporter, 0, 0, 1);
@@ -105,7 +105,7 @@ public class SystemTestsEntryPoint extends SystemTests {
 		Reporter reporter = setup("tests/Test01.xml", "tests/entrypoint/EntryPoint03-A.xml");
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 8, 0, 4);
 		checkInstrumentation(reporter, 24, 6, 1);
 		checkMarkerCoverage(reporter, 0, 0, 1);
@@ -117,7 +117,7 @@ public class SystemTestsEntryPoint extends SystemTests {
 		Reporter reporter = setup("tests/Test01.xml", "tests/entrypoint/EntryPoint03-B.xml");
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 8, 0, 4);
 		checkInstrumentation(reporter, 24, 6, 1);
 		checkMarkerCoverage(reporter, 0, 0, 1);