+++ src/test/java/za/ac/sun/cs/coastal/SystemTestsJars.java
@@ -20,7 +20,7 @@ public class SystemTestsJars extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 28, 6, 1);
 	}
@@ -35,7 +35,7 @@ public class SystemTestsJars extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 7);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 8, 0, 1);
 		checkInstrumentation(reporter, 49, 12, 1);
 	}
@@ -50,7 +50,7 @@ public class SystemTestsJars extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 28, 6, 1);
 	}
@@ -65,7 +65,7 @@ public class SystemTestsJars extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 28, 6, 1);
 	}