+++ src/test/java/za/ac/sun/cs/coastal/SystemTestsSimple.java
@@ -32,7 +32,7 @@ public class SystemTestsSimple extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 1);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 0, 0, 0);
 		checkInstrumentation(reporter, 7, 0, 1);
 	}
@@ -48,7 +48,7 @@ public class SystemTestsSimple extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 2);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 2, 0, 0);
 		checkInstrumentation(reporter, 14, 2, 1);
 	}
@@ -64,7 +64,7 @@ public class SystemTestsSimple extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 4);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 4, 0, 0);
 		checkInstrumentation(reporter, 28, 6, 1);
 	}
@@ -80,7 +80,7 @@ public class SystemTestsSimple extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 8);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 8, 0, 0);
 		checkInstrumentation(reporter, 56, 14, 1);
 	}
@@ -96,7 +96,7 @@ public class SystemTestsSimple extends SystemTests {
 		Reporter reporter = coastal.getReporter();
 		checkDivers(reporter, 1, 5);
 		checkSurfers(reporter, 0, 0);
-		checkStrategy(reporter, 1);
+		checkDFStrategy(reporter, 1);
 		checkPathTree(reporter, 6, 0, 1);
 		checkInstrumentation(reporter, 35, 8, 1);
 	}